restructure overlay
