+++ src/components/App.jsx
@@ -1,13 +1,11 @@
 import '../scss/main.scss'
 import React from 'react'
-import Dashboard from './Dashboard.jsx'
+import Layout from './Layout'
 
 class App extends React.Component {
   render () {
     return (
-      <div>
-        <Dashboard />
-      </div>
+      <Layout />
     )
   }
 }
